+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Adding Network information data
 - Put network data in DB
 - Update Json builder(add network data)
+- Adding load Avereage params
+- Adding cpu informatio
+- Adding version
 
 ## [0.1.9] - 2021-05-
 ### Changes
@@ -19,7 +22,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Adding some debug prints
 - Adding syslog
 - Adding network params
-- Adding load Avereage params
 
 ## [0.1.6] - 2021-05-06
 ### Changes